# src/main/resources/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Athleticore API","description":"API documentation for Athleticore Spring Boot application","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/session/":{"post":{"tags":["session-controller"],"operationId":"createSession","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}},"delete":{"tags":["session-controller"],"operationId":"deleteSession","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/api/clients/":{"get":{"tags":["client-controller"],"operationId":"getAllClients","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Client"}}}}}}},"post":{"tags":["client-controller"],"operationId":"addClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}}},"/api/session/{id}":{"patch":{"tags":["session-controller"],"operationId":"updateSession","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionFields"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/api/session/{idUser}":{"get":{"tags":["session-controller"],"operationId":"getSessionByUser","parameters":[{"name":"idUser","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/api/notifications/":{"get":{"tags":["notification-controller"],"operationId":"sendNotification","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/api/":{"get":{"tags":["main-controller"],"summary":"Get a greeting message","description":"Returns a simple greeting message","operationId":"getHomePage","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/clients/{id}":{"delete":{"tags":["client-controller"],"operationId":"deleteClientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"object"}},{"type":"string"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}}},"components":{"schemas":{"FullName":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"}}},"Schedule":{"type":"object"},"SessionDto":{"required":["category","date","name","sessionType","trainer"],"type":"object","properties":{"name":{"maxLength":12,"minLength":0,"type":"string"},"description":{"maxLength":60,"minLength":0,"type":"string"},"sessionType":{"type":"string","enum":["GROUP","INDIVIDUAL"]},"date":{"type":"string","format":"date"},"duration":{"type":"integer","format":"int32"},"trainer":{"$ref":"#/components/schemas/Trainer"},"difficulty":{"type":"string","enum":["LIGHT","NORMAL","HARD","PROF"]},"category":{"type":"string","enum":["YOGA","PILATES","BOXING","SWIMMING"]},"maxParticipants":{"type":"integer","format":"int32"}}},"Trainer":{"type":"object","properties":{"fullName":{"$ref":"#/components/schemas/FullName"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["CLIENT","TRAINER","MANAGER"]},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"schedule":{"$ref":"#/components/schemas/Schedule"}}},"ClientDto":{"type":"object","properties":{"fullName":{"$ref":"#/components/schemas/FullName"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"phoneNumber":{"type":"string"}}},"Client":{"type":"object","properties":{"fullName":{"$ref":"#/components/schemas/FullName"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["CLIENT","TRAINER","MANAGER"]},"id":{"type":"integer","format":"int64"},"phoneNumber":{"type":"string"},"notifications":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Notification"}}}},"Notification":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"message":{"type":"string"},"date":{"type":"string","format":"date"},"clients":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Client"}},"session":{"$ref":"#/components/schemas/Session"}}},"Session":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"sessionType":{"type":"string","enum":["GROUP","INDIVIDUAL"]},"date":{"type":"string","format":"date"},"duration":{"type":"integer","format":"int32"},"trainer":{"$ref":"#/components/schemas/Trainer"},"category":{"type":"string","enum":["YOGA","PILATES","BOXING","SWIMMING"]},"difficulty":{"type":"string","enum":["LIGHT","NORMAL","HARD","PROF"]},"maxParticipants":{"type":"integer","format":"int32"},"schedule":{"$ref":"#/components/schemas/Schedule"}}},"UpdateSessionFields":{"type":"object"}}}}